requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
starlette==0.46.2
itsdangerous==2.2.0
